# ENH: Refactor LoRA bnb layers for faster initialization